pyproject.toml
[project]
name = "codenames-embeddings-explorer"
version = "0.1.0"
description = "A Codenames hint generator using word embeddings"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "streamlit>=1.30.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]
